+++ includes/model/class-user.php
@@ -428,7 +428,16 @@ class User extends Actor {
 	 * @return array The alsoKnownAs.
 	 */
 	public function get_also_known_as() {
-		return \get_user_option( 'activitypub_also_known_as', $this->_id );
+		$also_known_as = array(
+			\add_query_arg( 'author', $this->_id, \home_url( '/' ) ),
+			$this->get_url(),
+			$this->get_alternate_url(),
+		);
+
+		// phpcs:ignore Universal.Operators.DisallowShortTernary.Found
+		$also_known_as = array_merge( $also_known_as, \get_user_option( 'activitypub_also_known_as', $this->_id ) ?: array() );
+
+		return array_unique( $also_known_as );
 	}
 
 	/**