+++ tests/includes/class-test-sanitize.php
@@ -57,6 +57,14 @@ class Test_Sanitize extends \WP_UnitTestCase {
 				),
 			),
 			'empty_array'                     => array( array(), array() ),
+			'unsupported'                     => array(
+				array(
+					'',
+					false,
+					null,
+				),
+				array(),
+			),
 		);
 	}